src/app/backend/requirements.txt
fastapi
uvicorn
langchain
openai
python-dotenv
pymupdf
